test: add new test to fit with controller parameter in Participant
such as
- Participants has two constructors which have `String/Player` as parameter, but InputView returning `List<String>`
- change directories for refactor
