--- src/test/kotlin/blackjack/model/ParticipantTest.kt
@@ -1,41 +1,0 @@
-package blackjack.model
-
-import blackjack.model.participant.Participants
-import blackjack.model.participant.Player
-import blackjack.model.participant.Dealer
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
-
-class ParticipantTest {
-    @Test
-    fun `Participants init with player and dealer, and recognize by name`() {
-        val name = "mina"
-        val participants = Participants(name)
-
-        assertThat(participants.contain(name)).isTrue()
-    }
-
-    @Test
-    fun `Participants init with player and dealer, and recognize by type of Participant`() {
-        val player = Player("name")
-        val participants = Participants(player)
-        val dealer = participants.getDealer()
-
-        assertThat(participants.contain(dealer)).isTrue()
-    }
-
-    @Test
-    fun `Participants init with players`() {
-        val participants = Participants("mina", "guri")
-
-        assertThat(participants.containsAll("mina", "guri")).isTrue()
-        assertThat(participants.containsAll("guri", "I am not Guri")).isFalse()
-        assertThat(participants[0].name).isEqualTo("mina")
-    }
-
-    @Test
-    fun `can recognize number of players`() {
-        val participants = Participants("mina", "guri")
-        assertThat(participants.numberOfPlayers()).isEqualTo(2)
-    }
-}
